requirements.txt
setuptools
torchmetrics
torchinfo
tf-keras
tensorflow
transformers
accelerate
datasets
einops
numpy==1.26.4
pandas
scikit-learn
scipy
wandb
peft
matplotlib
peft
biopython
seaborn
matplotlib
streamlit
rapidfuzz
gseapy
biotite
rapidfuzz